tests/wpt/web-platform-tests/credential-management/fedcm-revoke.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>FederatedCredential.revoke() promise resolution</title>
<link rel="author" title="Christian Biesinger" href="mailto:cbiesinger@chromium.org">
<link rel="help" href="https://wicg.github.io/FedCM/#browser-api-revocation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script type="module">
  import {fedcm_test} from './support/fedcm-helper.js';

  const provider = {
    url: "https://idp.example/",
    clientId: "1234",
    nonce: "5678"
  };

  fedcm_test(async (t, mock) => {
    mock.revokeReturn("kSuccess");
    await FederatedCredential.revoke("foo@bar.com", provider);
  }, "Successfully revoking a token should resolve the promise.");

  fedcm_test(async (t, mock) => {
    mock.revokeReturn("kError");
    const result = FederatedCredential.revoke("foo@bar.com", provider);
    return promise_rejects_dom(t, "NetworkError", result);
  }, "Error should reject the promise.");
</script>
